requirements.txt
# ============================================
# 📦 Python Dependencies
# Driver Drowsiness Detection System
# ============================================

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Database
mysql-connector-python==8.2.0

# GUI
customtkinter==5.2.1
Pillow==10.1.0

# Audio
pygame==2.5.2

# Environment Variables
python-dotenv==1.0.0

# Utilities
numpy==1.26.2

# Password Hashing
bcrypt==4.1.1

# For Head Pose Estimation
scipy==1.11.4
